requirements.txt
selenium==2.23.0
pyside
ghost.py
